% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_importFrom_fnctn_statments.R
\name{add_importFrom_fnctn_statments}
\alias{add_importFrom_fnctn_statments}
\title{Add @importFrom Statements Based on Script Dependencies}
\usage{
add_importFrom_fnctn_statments(script)
}
\arguments{
\item{script}{A character string specifying the path to the R script file.}
}
\value{
This function generates and prints \verb{@importFrom} statements for each
package detected in the script.
}
\description{
This function analyzes a given script to detect functions used in the
\verb{package::function} notation. It then generates \verb{@importFrom} statements
for each package detected, capturing the required dependencies for the script.
}
\examples{
ex_script <- "mtcars \%>\%
   dplyr::mutate(
     var_1 = DescTools::Quantile(disp, probs = .6),
     var_2 = str_glue('{disp} {hp}')
   ) \%>\%
   dplyr::arrange(disp)"
temp_script <- tempfile(fileext = ".R")
writeLines(ex_script, temp_script)
add_importFrom_fnctn_statments(temp_script)

}
