% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pkg_fnctn_definitions.R
\name{add_pkg_fnctn_definitions}
\alias{add_pkg_fnctn_definitions}
\title{Add Package Definitions to Functions in a Script}
\usage{
add_pkg_fnctn_definitions(script)
}
\arguments{
\item{script}{A character string specifying the path to the R script file.}
}
\description{
This function scans a provided R script for function calls that should have package qualifiers
based on the packages currently loaded in the session. If it finds any unqualified function calls,
it replaces them with the fully qualified notation \link{package}::\link{function}.
}
\details{
This function uses the session information to identify which packages are loaded and the functions
they contain. It then reads the script, identifies non-qualified function calls, and replaces them
with their qualified counterparts if a match is found in the loaded packages.
}
\examples{
\dontrun{
library(dplyr)
library(stringr)
library(DescTools)

ex_script <- "mtcars \%>\%
  mutate(
  var_1 = Quantile(disp, probs = .6),
         var_2 = str_glue('{disp} {hp}'))"

# Write the example script to a temporary file
temp_script <- tempfile(fileext = ".R")
writeLines(ex_script, temp_script)
add_pkg_fnctn_definitions(temp_script)
}
}
