% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remindMe_code_reorg_prompt.R
\name{remindMe_code_reorg_prompt}
\alias{remindMe_code_reorg_prompt}
\title{Print basic LLM prompt for to reorganize code.}
\usage{
remindMe_code_reorg_prompt()
}
\value{
None (prints to console).
}
\description{
Print basic LLM prompt for to reorganize code.
}
\examples{
\dontrun{
remindMe_packages_to_use()

}
}
