% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remindMe_function_doc_prompt.R
\name{remindMe_function_doc_prompt}
\alias{remindMe_function_doc_prompt}
\title{Print basic LLM prompt for function docummnetation.}
\usage{
remindMe_function_doc_prompt()
}
\value{
None (prints to console).
}
\description{
Print basic LLM prompt for function docummnetation.
}
\examples{
\dontrun{
remindMe_packages_to_use()

}
}
